refactor: move role definition utils
