+++ packages/amplify-graphql-auth-transformer/src/index.ts
@@ -1,6 +1,7 @@
 export * from './graphql-auth-transformer';
 export * from './utils/constants';
 export * from './utils/definitions';
+export { RoleDefinition, RolesByProvider } from './utils/role-definition';
 export { AccessControlMatrix } from './accesscontrol';
 export { getAuthDirectiveRules } from './utils';
 // No-op change to trigger publish
